+++ src/components/atoms/GiftFilter/index.tsx
@@ -22,6 +22,7 @@ export const GiftFilter = ({
   ...props
 }: Props) => {
   const [selected, setSelected] = useState<SelectedProps>(selectedProps)
+  const [friendsData, setFriendsData] = useRecoilState(friendState)
 
   useEffect(() => {
     setSelected(selectedProps)
@@ -49,6 +50,7 @@ export const GiftFilter = ({
       return item
     })
     setSelected(updatedBtn)
+    setFriendsData({ ...friendsData, presentRecommendFilterValue: '' })
 
     // 선택된 카테고리가 없으면 모든 카테고리를 선택한 상태로 업데이트
     const updatedTags = updatedBtn
@@ -70,7 +72,7 @@ export const GiftFilter = ({
       active: false,
     }))
     setSelected(updatedBtn)
-
+    setFriendsData({ ...friendsData, presentRecommendFilterValue: '' })
     // 모든 카테고리를 선택한 상태로 설정
     setSelectedTags(selectedProps.map(({ name, value }) => ({ name, value })))
   }
