+++ src/components/friends/PresentRecommend.tsx
@@ -28,8 +28,10 @@ import BottomSheet from '@components/atoms/BottomSheet'
 import useBottomSheet from '@libs/hooks/useBottomSheet'
 import { friendState } from '@libs/store/friend'
 import { CategoryNameType } from '@components/atoms/Category'
+import { useInfiniteQueries } from '@libs/hooks'
+import PresentItem from './PresentItem'
 
-type DataType = {
+export type DataType = {
   productId: number
   name: string
   brand: string
@@ -60,13 +62,10 @@ function PresentRecommend() {
   const [selectedProps, setSelectedProps] =
     useState<SelectedProps>(selectedPropsData)
   const [selectedTags, setSelectedTags] = useState<SelectedTag[]>([])
-  const [dataLoaded, setDataLoaded] = useState<boolean>(false)
-  const [products, setProducts] = useState<DataType[]>([])
   const [isSortOpen, setIsSortOpen] = useRecoilState<string>(isFilterTypeState)
   const selectedFilter = useRecoilValue(FilterState)
   const selectedPrice = useRecoilValue(PriceState)
   const friendStateData = useRecoilValue(friendState)
-  const [initialRender, setInitialRender] = useState(true)
 
   const selectedCategories = selectedTags.map((tag) => tag.value)
   const selectedCategoriesString = selectedCategories
@@ -78,67 +77,28 @@ function PresentRecommend() {
     .map((category) => `${category.value}`)
     .join(',')
 
-  const gotoSite = (siteUrl: string) => {
-    window.open(`${siteUrl}`)
-  }
-
-  const formatPrice = (price: number) => {
-    return price.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',')
-  }
-
-  const defaultImage = (smallCategory: string) => {
-    switch (smallCategory) {
-      case 'MAKEUP':
-        return makeupNull
-      case 'FRAGRANCE':
-        return fragranceNull
-      case 'CLOTHES':
-        return clothesNull
-      case 'DIGITAL_PRODUCT':
-        return digitalNull
-      case 'FASHION_PRODUCT':
-        return fashionNull
-      case 'BAG':
-        return bagNull
-      case 'ACCESSORY':
-        return accessoryNull
-      case 'COOKING':
-        return cookingNull
-      case 'EXERCISE':
-        return exerciseNull
-      default:
-        return makeupNull
-    }
-  }
   console.log(selectedProps)
   console.log('tags:', selectedTags)
 
-  useEffect(() => {
-    if (!initialRender) {
-      //처음 렌더링 시 전체 카테고리 호출 방지
-      FriendsApi.GET_PRESENT_RECOMMEND(
-        selectedTags.length > 0
-          ? selectedCategoriesString
-          : allCategoriesString,
-        selectedFilter.filterValue,
-        selectedPrice.priceValue,
-      ).then((response) => {
-        if (response.statusCode === 200) {
-          setProducts(response.data.content)
-          setDataLoaded(true)
-        } else {
-          setDataLoaded(false)
-        }
-      })
-    } else {
-      setInitialRender(false)
-    }
-  }, [
-    selectedTags,
-    selectedFilter.filter,
-    selectedPrice.price,
-    friendStateData.presentRecommendFilterValue,
-  ])
+  const { infiniteListElement, isEmpty } = useInfiniteQueries<DataType>(
+    [
+      'GET_PRESENT_RECOMMEND',
+      selectedTags,
+      selectedFilter.filter,
+      selectedPrice.price,
+      friendStateData.presentRecommendFilterValue,
+    ],
+    () =>
+      FriendsApi.GET_PRESENT_RECOMMEND({
+        smallCategory:
+          selectedTags.length > 0
+            ? selectedCategoriesString
+            : allCategoriesString,
+        priceOrder: selectedFilter.filterValue,
+        priceFilter: selectedPrice.priceValue,
+      }),
+    PresentItem,
+  )
 
   useEffect(() => {
     const selectedOption = selectedProps.find(
@@ -222,47 +182,7 @@ function PresentRecommend() {
           </RecommendFilter>
         </FilterItemWrap>
       </ContainerFix>
-      <Container>
-        {dataLoaded && (
-          <SortItemWrap>
-            {products.map((product: DataType, index: number) => (
-              <ItemDiv key={index} onClick={() => gotoSite(product.siteUrl)}>
-                <ItemImg
-                  imageUrl={
-                    product.imageUrl || defaultImage(product.smallCategory)
-                  }
-                />
-                <Text
-                  typo="Caption_12R"
-                  color="gray_400"
-                  children={product.brand}
-                />
-                <div
-                  style={{
-                    display: 'flex',
-                    marginBottom: '4px',
-                  }}
-                >
-                  <Text
-                    typo="Caption_12R"
-                    color="gray_100"
-                    children={
-                      product.productOption
-                        ? `${product.productOption}` + ' - ' + `${product.name}`
-                        : `${product.name}`
-                    }
-                  />
-                </div>
-                <Text
-                  typo="Mont_Caption_12B"
-                  color="gray_100"
-                  children={formatPrice(product.price)}
-                />
-              </ItemDiv>
-            ))}
-          </SortItemWrap>
-        )}
-      </Container>
+      <Container>{!isEmpty && infiniteListElement}</Container>
       <>
         <BottomSheet
           isexpanded={isBottomSheetOpen}
@@ -281,6 +201,7 @@ function PresentRecommend() {
 export default PresentRecommend
 
 const FilterWrapper = styled.div`
+  z-index: 1;
   width: 100%;
   overflow-x: auto;
   white-space: nowrap;
@@ -303,6 +224,7 @@ const Container = styled.div`
 const ContainerFix = styled.div`
   padding: 0px 24px 0 24px;
   position: sticky;
+  z-index: 1;
   top: 160px;
   width: 100%;
 `
@@ -327,28 +249,6 @@ const Margin = styled.div<{
 }>`
   width: ${({ widthProps }) => `${widthProps}px`};
   height: 100%;
-`
-const SortItemWrap = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  width: 100%;
-  row-gap: 52px;
-  column-gap: 12px;
-  justify-items: center;
-`
-const ItemDiv = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-`
-
-const ItemImg = styled.div<{ imageUrl: string }>`
-  width: 100%;
-  padding-bottom: 100%;
-  background-image: ${({ imageUrl }) => `url(${imageUrl})`};
-  background-size: cover;
-  border-radius: 8px;
-  margin-bottom: 8px;
 `
 const BottomSpacing = styled.div`
   width: 100%;
