+++ src/libs/hooks/useInfiniteQueries.tsx
@@ -24,7 +24,7 @@ export const useInfiniteQueries = <T,>(
     InfiniteResponse<T>,
     unknown
   >(queryKey, apiFunction, {
-    getNextPageParam: (lastPage) => lastPage.page + 1,
+    getNextPageParam: (lastPage) => lastPage.data.page + 1,
     ...options,
   })
 
@@ -32,12 +32,12 @@ export const useInfiniteQueries = <T,>(
     if (!data) return
 
     const lastPageIdx = data.pages.length - 1
-    const hasNext = data.pages[lastPageIdx].hasNext
+    const hasNext = data.pages[lastPageIdx].data.hasNext
     if (hasNext && inView) fetchNextPage()
   }, [inView])
-
-  const listElement = data?.pages.map(({ content }) =>
-    content.map((item, idx) => (
+  console.log(data)
+  const listElement = data?.pages.map(({ data }) =>
+    data.content.map((item: any, idx: any) => (
       <ListItem {...item} className={`item-${idx}`} key={`item-${idx}`} />
     )),
   )
@@ -49,7 +49,7 @@ export const useInfiniteQueries = <T,>(
     />
   )
 
-  const isEmpty = data?.pages[0].content.length === 1
+  const isEmpty = data?.pages[0].data.content.length === 1
 
   return {
     infiniteListElement: (
@@ -63,6 +63,10 @@ export const useInfiniteQueries = <T,>(
 }
 
 const ListElementContainer = styled.div`
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  row-gap: 52px;
+  column-gap: 12px;
   @keyframes fadeIn {
     from {
       opacity: 0;
@@ -109,14 +113,20 @@ const ListElementContainer = styled.div`
 `
 
 export interface InfiniteResponse<T> {
-  content: T[]
-  page: number
-  size: number
-  hasNext: boolean
+  data: {
+    content: T[]
+    page: number
+    size: number
+    hasNext: boolean
+  }
+  statusCode: number
+  success: boolean
 }
 
 export interface InfiniteRequest {
-  questionId: number
+  smallCategory: string
+  priceOrder: string
+  priceFilter: string
   pageParam?: number
   size?: number
   sort?: 'asc' | 'desc'
